% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_function.R
\name{anova_function}
\alias{anova_function}
\title{Creates an ANOVA and makes plots}
\usage{
anova_function(data, response, predictor)
}
\arguments{
\item{data}{your dataframe that you want to process}

\item{response}{the variable that you think is dependent on the predictor}

\item{predictor}{the variable that explains the relationship}
}
\value{
summary of the ANOVA results and plots of the data
}
\description{
Creates an ANOVA and makes plots
}
