% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_process_field_function.R
\name{as.ID}
\alias{as.ID}
\title{as.ID
specify unique primary key in table, and check uniqueness}
\usage{
as.ID(dataVector)
}
\arguments{
\item{dataVector}{value vector}
}
\value{
boolean
}
\description{
as.ID
specify unique primary key in table, and check uniqueness
}
