% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_dc_neighborhoods.R
\name{assign_neighborhood}
\alias{assign_neighborhood}
\title{Assigns neighborhood labels to properties based on latitude and longitude}
\usage{
assign_neighborhood(
  df,
  geojson_path = system.file("extdata", "neighbourhoods_cleaned.geojson", package =
    "DC_property_data")
)
}
\arguments{
\item{df}{A dataframe containing `latitude` and `longitude` columns.}

\item{geojson_path}{The file path to the GeoJSON file with neighborhood boundaries.}
}
\value{
A dataframe with an additional column `neighborhood` indicating the assigned neighborhood.
}
\description{
This function takes a dataframe with latitude and longitude columns and assigns
each property a neighborhood based on a provided GeoJSON file. Properties
outside any defined neighborhood are labeled as "Unknown".
}
\examples{
\dontrun{
df <- data.frame(latitude = c(38.9, 38.85), longitude = c(-77.0, -76.95))
geojson_path <- "neighbourhoods_cleaned.geojson"
df <- assign_neighborhood(df, geojson_path)
}
}
