% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_chi.R
\name{assumptions_chi}
\alias{assumptions_chi}
\title{Assumptions for Chi-Squared Test}
\usage{
assumptions_chi(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns list.
}
\description{
Tabular assumptions for chi-squared test (expected values).
}
