% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_t.R
\name{assumptions_t}
\alias{assumptions_t}
\title{Assumptions for T-test}
\usage{
assumptions_t(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns assumptions as a list.
}
\description{
Provides the t-test graphical assumptions.
}
