% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Posterior support bar plot}
\usage{
bar_plot(df, column, title, x_title, y_title)
}
\arguments{
\item{df}{the data frame you want to pull data from}

\item{column}{the column you want to plot}

\item{title}{title of plot}

\item{x_title}{title of x axis}

\item{y_title}{title of y axis}
}
\value{
A bar graph displaying the frequency of column selected
}
\description{
Posterior support bar plot
}
