% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{bigen}
\alias{bigen}
\title{bigen}
\usage{
bigen(p, n)
}
\arguments{
\item{p}{is the probability}

\item{n}{is the number of times you want to repeat the bernoulli trial.}
}
\description{
A binomial random variable can be seen as the result of repeated Bernoulli Trials. This function gives you a number which follows binomial distribution. It uses bernoulli distribution function (bergen (p)) to make binomial numbers.
}
\seealso{
dugen(), cugen(), brgen(), gegen(), expgen(), gagen(), pogen(), nogen()
}
\author{
AmirHossein Motameni
}
