% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_process_field_function.R
\name{boolean}
\alias{boolean}
\title{boolean
check if vector values are boolean}
\usage{
boolean(dataVector)
}
\arguments{
\item{dataVector}{value vector}
}
\value{
boolean
}
\description{
boolean
check if vector values are boolean
}
