% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{box_cox_likelihood}
\alias{box_cox_likelihood}
\title{Boxcox likelihood function. Used internally to determine the best-performing
transformation from a search space of lambdas.}
\usage{
box_cox_likelihood(x, lambda)
}
\arguments{
\item{x}{One dimensional dataset to used for likelihood calculation}

\item{lambda}{One dimensional container of lambdas for testing.}
}
\description{
Boxcox likelihood function. Used internally to determine the best-performing
transformation from a search space of lambdas.
}
\examples{

# Load iris dataset 
data('iris')

# Plot likelihood across a range of lambdas
lambdas = seq(-5,5,0.025)
plot(lambdas, box_cox_likelihood(iris[1:25,1], lambdas)

}
