% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxR.R
\name{boxR}
\alias{boxR}
\title{boxR}
\usage{
boxR(data, x, y, title, xname, yname, points, savename)
}
\arguments{
\item{data}{data that will be plotted}

\item{x}{x axis variable}

\item{y}{y axis variable}

\item{title}{plot title}

\item{xname}{x axis name}

\item{yname}{y axis name}

\item{points}{whether or not you want individual data points on your box plots}

\item{savename}{what to save the plot as}
}
\value{
plots either a boxplot or dotplot
}
\description{
boxR
}
