% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{brgen}
\alias{brgen}
\title{brgen}
\usage{
brgen(p)
}
\arguments{
\item{p}{is the probability}
}
\description{
Bernoulli distribution is one the most simple as well as famous distributions. In each Bernoulli trial there are two possible outcomes, namely success or failure, for the probability of success equal to p, the probability of failure is 1 ??? p.
}
\seealso{
dugen(), cugen(), gegen(), expgen(), gagen(), pogen(), nogen(), bigen()
}
\author{
Reza Mousapour
}
