% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkR.R
\name{chunkR}
\alias{chunkR}
\title{chunkR}
\usage{
chunkR(data, group, selection, sum.name)
}
\arguments{
\item{data}{the data in which you want to subset}

\item{group}{what you want to group the data by}

\item{selection}{what you want to select from the data}

\item{sum.name}{name of summarize column}
}
\value{
saves a chunk of a larger dataset as a new smaller dataset
}
\description{
chunkR
}
