% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciekawa funkcja.R
\name{ciekawafunkcja}
\alias{ciekawafunkcja}
\title{Ciekawa funkcja}
\usage{
ciekawafunkcja(x)
}
\arguments{
\item{x}{An integer value}
}
\description{
Jest to pierwsza funkcja stworzona przez Wojtka w tym programie, wtedy jeszcze uważał on,
że jest ona ciekawa
}
