% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characater_indicator_plot.R
\name{clean_data}
\alias{clean_data}
\title{Bar plot of Character Indicator}
\usage{
clean_data(filepath, columnname)
}
\arguments{
\item{filepath}{the path to the file you want to process}

\item{columnname}{Character Indicator column name}
}
\value{
Plot of Posterior Support by Character Indicator
}
\description{
Bar plot of Character Indicator
}
