% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanR.R
\name{cleanR}
\alias{cleanR}
\title{cleanR}
\usage{
cleanR(data, clean_names)
}
\arguments{
\item{data}{the data you wish to clean}

\item{clean_names}{whether or not you want to remove bad naming conventions like spaces}
}
\value{
saves clean data set with no NA values and clean naming conventions
}
\description{
cleanR
}
