% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files.R
\name{combine_logs}
\alias{combine_logs}
\title{Combine Two files with Source Column}
\usage{
combine_logs(df1, df2, name1, name2)
}
\arguments{
\item{df2}{The second data frame to combine.}

\item{name1}{The name to assign as the source for the first data frame.}

\item{name2}{The name to assign as the source for the second data frame.}
}
\value{
A combined data frame with an additional "source" column indicating the origin of each row.
}
\description{
@param df1 The first data frame to combine.
}
