% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion_chi.R
\name{conclusion_chi}
\alias{conclusion_chi}
\title{Conclusion for Chi-Squared test}
\usage{
conclusion_chi(object)
}
\arguments{
\item{object}{List output "mychi_chi" from the chi_chi function.}
}
\value{
Returns decision as a list.
}
\description{
Summarises the result for the chi-squared test.
}
