% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q1.R
\name{conclusion_linear}
\alias{conclusion_linear}
\title{Test if there is a linear relationship}
\usage{
conclusion_linear(myfit1)
}
\arguments{
\item{myfit1}{the beta and p value from linear regression model}
}
\value{
A statistic test with the hypothesis, tests assumptions, test statistics, a decision and conclusion.
}
\description{
Conducts a linear test to see if there is a relationship between heights and weight.
}
