% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion_lm.R
\name{conclusion_lm}
\alias{conclusion_lm}
\title{Conclusion for Linear Regression}
\usage{
conclusion_lm(object)
}
\arguments{
\item{object}{List output "myfit_lm" from the chi_chi function.}
}
\value{
Returns conclusion as a list.
}
\description{
Summarises the result for the linear regression model.
}
