% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion_t.R
\name{conclusion_t}
\alias{conclusion_t}
\title{Conclusion for T-test}
\usage{
conclusion_t(object)
}
\arguments{
\item{object}{List output "myttest_t" from the ttest_t function.}
}
\value{
Returns conclusion as a list.
}
\description{
Summarises the result for the t-test.
}
