% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{cugen}
\alias{cugen}
\title{cugen}
\usage{
cugen()
}
\description{
generate uniform random numbers between 0 and 1, using function dugen(a,b) which taking the two integer numbers as input and generates random numbers uniformly distributed between two integer inputs.
}
\seealso{
dugen(), cugen(), brgen(), bigen(), gegen(), expgen(), gagen(), pogen(), nogen()
related functions: dugen
}
\author{
MohammadAmin Salarkia
}
