% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_chi.R
\name{decision_chi}
\alias{decision_chi}
\title{Decision for Chi-Squared test.}
\usage{
decision_chi(object)
}
\arguments{
\item{object}{List output "mychi_chi" from the chi_chi function.}
}
\value{
Returns decision as a list.
}
\description{
Provides the "decision" on which hypothesis to reject/accept.
}
