% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_lm.R
\name{decision_lm}
\alias{decision_lm}
\title{Decision for Linear Regression model.}
\usage{
decision_lm(object)
}
\arguments{
\item{object}{List output "myfit_lm" from the chi_chi function.}
}
\value{
Returns decision as a list.
}
\description{
Provides the "decision" on which hypothesis to reject/accept.
}
