% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_t.R
\name{decision_t}
\alias{decision_t}
\title{Decision for T-test.}
\usage{
decision_t(object)
}
\arguments{
\item{object}{List output "myttest_t" from the chi_chi function.}
}
\value{
Returns decision as a list.
}
\description{
Provides the "decision" on which hypothesis to reject/accept.
}
