% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violation-Constraint-Management.R
\name{delete}
\alias{delete}
\title{delete
Delete data at indice when foreing key constraint were violated}
\usage{
delete(tables, indice, fk_field, fk_table, pk_table, namePrefix)
}
\arguments{
\item{tables}{list of data table}

\item{indice}{row indice in fk_table}

\item{fk_field}{column name on which foreign key constraint is checked}

\item{fk_table}{name of data frame on which foreign key constraint is checked}

\item{pk_table}{name of data frame corresponding to table checking}

\item{namePrefix}{prefix flag column when flag like function is called (unused in this f)}
}
\description{
delete
Delete data at indice when foreing key constraint were violated
}
