\name{demodata}
\alias{demodata}
\docType{data}
\title{
Mean and SD
}
\description{
Gives the mean and standard deviation of 100 random numbers
}
\usage{data(demodata)}
\format{
  The format is:
 num [1:100] 0.542 1 -0.244 -1.132 0.666 ...
}
\details{ }
\source{
Dr. LEE 550.400 class
}
\references{
}
\examples{

}
\keyword{datasets}
