% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{density plot}
\usage{
density_plot(df, x_column1, x_column2, plot_title, x_title, y_title)
}
\arguments{
\item{df}{the data frame you want to pull data from}

\item{x_column1}{first column you want to plot}

\item{x_column2}{second column you want to plot}

\item{x_title}{title of x axis}

\item{y_title}{title of y axis}

\item{title}{title of plot}
}
\value{
A density plot displaying the frequency of column selected
}
\description{
density plot
}
