% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example Data}
\format{
A list with 3 elements:
\describe{
\item{time}{Numeric vector.}
\item{group}{Numeric vector.}
\item{CountData}{Data frame with 500 rows and 36 variables.}
}
}
\source{
Generated for demonstration purposes
}
\usage{
data(ExampleData)
}
\description{
This dataset contains information on time, group, and count data for analysis purposes.
It is a list with 3 elements:
\describe{
\item{time}{A numeric vector of length 36 representing time points.}
\item{group}{A numeric vector of length 36 representing group IDs.}
\item{CountData}{A data frame with 500 rows and 36 variables containing count data.}
}
}
\keyword{datasets}
