% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_registry_qc_doc.R
\name{fill_registry_qc_doc}
\alias{fill_registry_qc_doc}
\title{fill_registry_qc_doc}
\usage{
fill_registry_qc_doc(
  sp_qc_doc_path,
  results_folder,
  final_name,
  analyst_initials
)
}
\arguments{
\item{sp_qc_doc_path}{path to the registry QC document file}

\item{results_folder}{local results or reports folder to save the filled in QC document}

\item{final_name}{the final QC document file name}

\item{analyst_initials}{Initials of the analyst}
}
\value{
returns a list with the raw filled in result table and the flextable as well as the saved QC document
}
\description{
Function that fills in the Registry QC document for monthly and quarterly reports
}
\examples{
# fill_registry_qc_doc("C:/Users/NicoleFergestrom/Corrona LLC/Biostat and Epi Team Site - Biostat QC/Biostat Registry - Monthly and Quarterly Report QC Checklist.docx",
# results_folder = "C:/Users/NicoleFergestrom/OneDrive - Corrona LLC/OneDrive Corevitas/Documents/scrap/corProject_test",
# final_name = "Filled QC Registry QC Checklist",
# analyst_initials = "NF")

}
