% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q1.R
\name{fit_linear}
\alias{fit_linear}
\title{Test if there is a linear relationship}
\usage{
fit_linear(file_name)
}
\arguments{
\item{file_name}{The name of the excel file that will be read}
}
\value{
A statistic test with the hypothesis, tests assumptions, test statistics, a decision and conclusion.
}
\description{
Conducts a linear test to see if there is a relationship between heights and weight.
}
