% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violation-Constraint-Management.R
\name{flag}
\alias{flag}
\title{flag
flag violated constraint data into new annotation column in fk_table (named <prefixName>_<pk_table>)}
\usage{
flag(tables, indice, fk_field, fk_table, pk_table, namePrefix)
}
\arguments{
\item{tables}{list of data table}

\item{indice}{row indice in fk_table}

\item{fk_field}{column name on which foreign key constraint is checked}

\item{fk_table}{name of data frame on which foreign key constraint is checked}

\item{pk_table}{name of data frame corresponding to table checking}

\item{namePrefix}{prefix flag column when flag like function is called (unused in this f)}
}
\description{
flag
flag violated constraint data into new annotation column in fk_table (named <prefixName>_<pk_table>)
}
