% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint-Management.R
\name{foreignKeyConstraint}
\alias{foreignKeyConstraint}
\title{foreignKeyConstraint
check foreign key constraint and apply action defined by user}
\usage{
foreignKeyConstraint(tables, fk_field, fk_table, pk_field, pk_table,
  action)
}
\arguments{
\item{tables}{list of data table}

\item{fk_field}{column name on which foreign key constraint is checked}

\item{fk_table}{name of data frame on which foreign key constraint is checked}

\item{pk_field}{name of field corresponding to the foreign key check}

\item{pk_table}{name of data frame corresponding to table checking}

\item{action}{function name that apply certain action when contraint is violated, by default three function name ("delete","flag","delete_flag") on}
}
\description{
foreignKeyConstraint
check foreign key constraint and apply action defined by user
}
