% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstpaiR.R
\name{FstpaiR}
\alias{FstpaiR}
\title{FstpaiR}
\usage{
FstpaiR(data, pcoa)
}
\arguments{
\item{data}{the data in which to analyse}

\item{pcoa}{whether or not you want a pcoa plot of the data}
}
\value{
prints a matrix of Fst values and a pcoa plot if specified
}
\description{
FstpaiR
}
