% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{gagen}
\alias{gagen}
\title{gagen}
\usage{
gagen(theta, k)
}
\arguments{
\item{k}{is an integer}

\item{Theta}{is a float number as the parameter of the underlying exponential distribution}
}
\description{
Summation of k i.i.d exponential random variables leads to a gamma distributed random variable.
}
\seealso{
dugen(), cugen(), brgen(), bigen(), gegen(), expgen(), gagen(), pogen(), nogen()
related functions: expgen
}
\author{
MohammadAmin Salarkia
}
