% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{gegen}
\alias{gegen}
\title{gegen}
\usage{
gegen(p)
}
\arguments{
\item{p}{is the probability}
}
\description{
The number of failures in Bernoulli trials, between two wins, follows the geometric distributions.
}
\seealso{
dugen(), cugen(), brgen(), bigen(), gegen(), expgen(), gagen(), pogen(), nogen()
related functions: brgen
}
\author{
Reza Mousapour
}
