% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_process_field_function.R
\name{generatedKey}
\alias{generatedKey}
\title{generatedKey
generate character key on specified dataframe subset}
\usage{
generatedKey(dataFrame)
}
\arguments{
\item{dataFrame}{column subset data frame on which identifiers should be generated}
}
\value{
vector
}
\description{
generatedKey
generate character key on specified dataframe subset
}
