% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_process_field_function.R
\name{generateWirdNomenclatureTable}
\alias{generateWirdNomenclatureTable}
\title{generateWirdNomenclatureTable
generate table nomenclature from extended table when length(unique(id) != length(unique(label))}
\usage{
generateWirdNomenclatureTable(dataFrame)
}
\arguments{
\item{dataFrame}{column subset data frame on which identifiers should be generated}
}
\value{
vector
}
\description{
generateWirdNomenclatureTable
generate table nomenclature from extended table when length(unique(id) != length(unique(label))
}
