% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_specification.R
\name{get_sources_file}
\alias{get_sources_file}
\title{get_sources_file
generate csv file and data frame compliant with wrangle.Files parameters}
\usage{
get_sources_file(folder = NA)
}
\arguments{
\item{file}{file_path to generate source_file}
}
\value{
data.frame
}
\description{
get_sources_file
generate csv file and data frame compliant with wrangle.Files parameters
}
