% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_specification.R
\name{get_wrangler_parameter_file}
\alias{get_wrangler_parameter_file}
\title{get_wrangler_parameter_file
generate csv file and data frame compliant with wrangle.Files parameters}
\usage{
get_wrangler_parameter_file(folder = NA)
}
\arguments{
\item{folder}{folder location in which csv  wrangler_parameter_file should be generated}
}
\value{
data.frame
}
\description{
get_wrangler_parameter_file
generate csv file and data frame compliant with wrangle.Files parameters
}
