\name{graph_mean}
\alias{graph_mean}
\title{Graph Mean}
\description{It takes the data generated by the mean_year function and plots it in a column graph.}
\usage{
graph_mean(x)
}
\arguments{
  \item{x}{Dataframe generated by mean_year}
}
\value{
A column graph with data generated by mean_year
}
\examples{
graph_mean(dataframe)
graph_mean(mean_year(2020:2024))
}