% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutselli_function.R
\name{gutselli_function}
\alias{gutselli_function}
\title{Finds average and returns text if size is too small}
\usage{
gutselli_function(filepath, column, number)
}
\arguments{
\item{filepath}{The path to the tab-delimited file you want to process}

\item{column}{The column in the data you want to find the average of}
}
\value{
"Too small" if average is less than it should be
}
\description{
Finds average and returns text if size is too small
}
