% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{hello_world}
\usage{
hello_world(your_name = "ExampleNameium")
}
\arguments{
\item{your_name}{The name to be displayed (default is ExampleNameium)}
}
\description{
hello_world
}
