% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_chi.R
\name{hypothesis_chi}
\alias{hypothesis_chi}
\title{Hypotheses for Chi-Squared Test}
\usage{
hypothesis_chi(data1)
}
\arguments{
\item{data1}{A dataframe.Use "project2022" dataset for this.}
}
\value{
Returns hypotheses as a list.
}
\description{
Provides the null and alterantive hypotheses.
}
