% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_lm.R
\name{hypothesis_lm}
\alias{hypothesis_lm}
\title{Hypotheses for Linear Regression}
\usage{
hypothesis_lm(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns a list.
}
\description{
Provides the null and alternative hypotheses.
}
