% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_t.R
\name{hypothesis_t}
\alias{hypothesis_t}
\title{Hypotheses for T-test}
\usage{
hypothesis_t(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns hypotheses as a list.
}
\description{
Provides the hypotheses for the t-test.
}
