% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWrangle.R
\name{importTableFromSource}
\alias{importTableFromSource}
\title{importTableFromSource
Import data table list into R}
\usage{
importTableFromSource(parameter, useAll = FALSE)
}
\arguments{
\item{sources_file}{parameter file to source data (see source_format)}
}
\value{
table list
}
\description{
importTableFromSource
Import data table list into R
}
