% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWrangle.R
\name{importWranglerParameter}
\alias{importWranglerParameter}
\title{importWranglerParameter
Get wrangler parameter data from source file}
\usage{
importWranglerParameter(wrangler_parameter)
}
\arguments{
\item{wrangler_parameter}{wrangler parameter file to transform data (excel or csv-comma)}
}
\value{
wrangler data frame
}
\description{
importWranglerParameter
Get wrangler parameter data from source file
}
