% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_function.R
\name{inches}
\alias{inches}
\title{Converts column from centimeters to inches}
\usage{
inches(filepath, column)
}
\arguments{
\item{filepath}{The path to the tab-delimited file you want to process}

\item{column}{The column in the data you want to convert}
}
\value{
converted data frame column
}
\description{
Converts column from centimeters to inches
}
