% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{info Print information about the available functions in this package}
\usage{
info()
}
\description{
info Print information about the available functions in this package
}
\examples{
project::info()
}
