% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwadrat.R
\name{kwadrat}
\alias{kwadrat}
\title{Kwadrat liczby}
\usage{
kwadrat(x)
}
\arguments{
\item{x}{A numeric value to square}
}
\description{
Funkcja ta dla danej wartosc x wyznacza jej kwadrat.
}
\examples{
kwadrat(5)=25
kwadrat(6)=36
}
