\name{letternumber}
\alias{letternumber}
\title{Letter Number Converter}
\description{Converts a string of letters into numbers. Each number is associated with its position in the alphabet.}
\usage{letternumber(phrase)}
\arguments{
  \item{phrase}{The phrase to be converted.}
}
\value{A numerical representation of a given alphabetical phrase.}
\examples{letternumber(phrase = "Deer")}