% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_function.R
\name{log_transform}
\alias{log_transform}
\title{Log10 transform data column}
\usage{
log_transform(filepath, column)
}
\arguments{
\item{filepath}{The path to the tab-delimited file you want to process}

\item{column}{The column in the data you want to transform}
}
\value{
a new column with the transformed data
}
\description{
Log10 transform data column
}
