% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{logistic_model}
\alias{logistic_model}
\title{logistic model function}
\usage{
logistic_model(data = heart, predictors)
}
\arguments{
\item{predictors}{possible predictors in heart attack data. Please use the original "heart" data variable name.}
}
\value{
summary of logistic model
}
\description{
Fit a logistic regression model to find a heart attack probability. Please refer to "heart" data document for variable names.
}
\examples{
logistic_model(data=heart, c("age","cp"))
logistic_model(c("age","cp"))
}
