% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxr.R
\name{maxr}
\alias{maxr}
\title{Removes NA values from a data column and calculates the numerical maximum.}
\usage{
maxr(path, column)
}
\arguments{
\item{path}{The file path to the comma-separated file.}

\item{column}{The column that you wish to calculate a maximum for.}
}
\value{
Prints a numerical maximum
}
\description{
Removes NA values from a data column and calculates the numerical maximum.
}
