% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanr.R
\name{meanr}
\alias{meanr}
\title{Removes NA values from a data column and calculates the mean.}
\usage{
meanr(path, column)
}
\arguments{
\item{path}{The path to the tab-delimited file you to process.}

\item{column}{The column of data that you wish to calculate a mean for.}
}
\value{
Prints a numerical mean.
}
\description{
Removes NA values from a data column and calculates the mean.
}
