% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rick_Wk_6_function.R
\name{misspell_swap2}
\alias{misspell_swap2}
\title{Misspell generator}
\usage{
misspell_swap2(word, start_pos, split_num)
}
\arguments{
\item{word}{A word to misspell}

\item{start_pos}{The position of the first letter you'd like swapped}

\item{split_num}{The distance between the next the first and second letter you'd like swapped}
}
\value{
A misspelled word
}
\description{
Misspell generator
}
\examples{
misspell_swap2("test", 1,2)
}
