% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.R
\name{mmf}
\alias{mmf}
\title{mmf}
\usage{
mmf(thetahat = NULL, thetahatses = NULL, denscomp = NULL,
  cdfband = NULL)
}
\arguments{
\item{thetahat}{Vector of estimates.}

\item{thetahatses}{Vector containing standard error of estimates.}

\item{denscomp}{ggplot graph object comparing distribution with estimated parameters to sample.}

\item{cdfband}{ggplot graph object with empirical cdf and an enclosing Kolmogorov-Smirnov confidence band.}
}
\description{
Output object containing estimates and other convenience functionality.
}

