% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfoptim.R, R/mvnpdfsmart.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart}
\title{Title}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{is a matrix containing the observations in columns}

\item{mean}{mean vector of the multivariate gaussian distribution}

\item{varcovM}{variance covariance matrix}

\item{Log}{logical parameter (default TRUE)}
}
\value{
a list of the results
}
\description{
Title
}
\examples{
X <- matrix(c(-0.5, 1.5, 0, 1, -1, 1), nrow = 2)
}
