% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_new_function_bernice.R
\name{my_new_function}
\alias{my_new_function}
\title{Takes dataset and gives us mean value between two variables}
\usage{
my_new_function(data, var1, var2)
}
\arguments{
\item{data}{Any input data}

\item{var1}{Name of the first variable}

\item{var2}{Name of the second variable}
}
\value{
the variable mean
}
\description{
Takes dataset and gives us mean value between two variables
}
\examples{
project::my_new_function(cars, speed, dist)
}
