% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mywrapper_chi.R
\name{mywrapper_chi}
\alias{mywrapper_chi}
\title{Wrapper Function for Chi-Squared test}
\usage{
mywrapper_chi(project2022)
}
\arguments{
\item{project2022}{dataset.}
}
\value{
Returns a series of cat/plots/tables.
}
\description{
Calls the helper functions to provide output for the test.
}
