% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mywrapper_lm.R
\name{mywrapper_lm}
\alias{mywrapper_lm}
\title{Wrapper Function for Linear Regression Model}
\usage{
mywrapper_lm(project2022)
}
\arguments{
\item{project2022}{dataset.}
}
\value{
Returns a series of cat/plots/tables.
}
\description{
Calls the helper functions to provide output for the test.
}
