% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mywrapper_t.R
\name{mywrapper_t}
\alias{mywrapper_t}
\title{Wrapper Function for T-test}
\usage{
mywrapper_t(project2022)
}
\arguments{
\item{project2022}{dataset.}
}
\value{
Returns a series of cat/plots/tables.
}
\description{
Calls the helper functions to provide output for the test.
}
