% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{nogen}
\alias{nogen}
\title{nogen}
\usage{
nogen(u, s)
}
\arguments{
\item{u}{is the mean}

\item{s}{is the variance}
}
\description{
This function gives you a number which follows normal distribution. It uses poisson distribution function (pogen (10, 10)) to make normal distributed numbers. It is calculated by (x * sqrt(s) / 10) + ( u - 10 * sqrt(s))
}
\seealso{
dugen(), cugen(), brgen(), bigen(), gegen(), expgen(), gagen(), pogen()
}
\author{
Ali Ehteshami
}
