% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Given a dataframe, assess normality of each numeric variable.
When a variable exhibits non-normal behavior, attempt to
transform to normality.}
\usage{
normal(dataset, significance = 5, boxcox_lambda_search = seq(-5, 5, 0.01))
}
\arguments{
\item{dataset}{Validate normality of each variable within this dataset}

\item{significance}{Will attempt to verify each variable's normality to within
this significance}

\item{boxcox_lambda_search}{When the significance test fails, will attempt to
find the lambda from within this range which most effectively transforms
the data to near-normality}
}
\description{
Given a dataframe, assess normality of each numeric variable.
When a variable exhibits non-normal behavior, attempt to
transform to normality.
}
\examples{
data('iris')
irisnorm = normal(iris[1:25,])
}
