% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normailitycheckR.R
\name{normalitycheckR}
\alias{normalitycheckR}
\title{normalitycheckR}
\usage{
normalitycheckR(data)
}
\arguments{
\item{data}{the dataset or individual data column in which you want to run normality tests on}
}
\value{
prints multiple statistical test reports for shaprio.wilk, kolmogorov.smirnov, and anderson.darling tests
}
\description{
normalitycheckR
}
