% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normify.normal}
\alias{normify.normal}
\title{Return "normified" dataset from a preallocated normal object}
\usage{
\method{normify}{normal}(object, ...)
}
\arguments{
\item{data}{Dataset to normify}
}
\description{
Return "normified" dataset from a preallocated normal object
}
