% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAPlotR.R
\name{PCAplotR}
\alias{PCAplotR}
\title{PCAplotR}
\usage{
PCAplotR(data, loadings, clusters)
}
\arguments{
\item{data}{the data used to make the plot}

\item{loadings}{whether or not you want the data displayed with loadings}

\item{clusters}{whether or not you want data displayed as clusters}
}
\value{
plots a PCA with loadings or clusters
}
\description{
PCAplotR
}
