% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmmf.R
\name{plot.mmf}
\alias{plot.mmf}
\title{plot}
\usage{
\method{plot}{mmf}(x, ...)
}
\arguments{
\item{x}{mmf to handle}

\item{...}{unused}
}
\description{
This function should be called by plot(object) and displays two plots side by side.
 The left plot is an empirical cdf with a confidence bound in red and blue around it.
 The right plot a graphing of the parametric and non-parametric data.
}
\examples{
 plot(testbeta())
}

