% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{plot.normal}
\alias{plot.normal}
\title{Generate QQ plots for each variable in the dataset.}
\usage{
\method{plot}{normal}(object, ...)
}
\arguments{
\item{object}{normal object}
}
\description{
Generate QQ plots for each variable in the dataset.
}
\examples{

# Load iris dataset 
data('iris')

plot(normal(iris[1:25,], significance=5))

}
