% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{plot.normal_assessment}
\alias{plot.normal_assessment}
\title{Create QQ plot from a normality_assessment object; called implicitely.}
\usage{
\method{plot}{normal_assessment}(object, ...)
}
\arguments{
\item{object}{normal_assessment object}
}
\description{
Create QQ plot from a normality_assessment object; called implicitely.
}
\examples{

# Load iris dataset 
data('iris')

plot(normal_assessment(iris[1:25,1], significance=5))

}
