% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Automatically creates plots based on the provided data}
\usage{
plot(data, x, y = NULL)
}
\arguments{
\item{data}{A data frame to plot.}

\item{x}{The name of the variable on the x-axis.}

\item{y}{(Optional) The name of the variable on the y-axis.}
}
\value{
A ggplot2 object with the created plot.
}
\description{
This function creates histograms, bar plots, or scatter plots depending on the provided variables.
}
\examples{
data(mtcars)
p <- plot(mtcars, "mpg", "hp")
}
