% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generating.R
\name{plot_generating}
\alias{plot_generating}
\title{Visualizing the heart data set}
\usage{
plot_generating(data = heart, variable)
}
\arguments{
\item{data}{using an inner dataset "heart"}

\item{variable}{choose the variable to visualize}
}
\description{
Plot a boxplot or barplot of variables for inner dataset "heart". Look at the data set document for variable input
}
\examples{
plot_generating(data=heart, variable="cp")


}
