% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plots}
\alias{plots}
\title{plots}
\usage{
plots(object)
}
\arguments{
\item{object}{The seurat object created by readSNV}
}
\description{
This function takes one argument, which is the seurat object representing the SNV data. The function returns a UMAP and gene by nuclei plot of the sample.
}
