% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotting}
\alias{plotting}
\title{A simple point plot of data}
\usage{
plotting(data, x, y, title = "title", x_lab = "x-lab", y_lab = "y-lab")
}
\arguments{
\item{data}{the data you want to plot}

\item{x}{the variable you want on the x-axis}

\item{y}{the variable you want on the y-axis}
}
\description{
A simple point plot of data
}
