% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{pogen}
\alias{pogen}
\title{pogen}
\usage{
pogen(lambda, t)
}
\arguments{
\item{lambda}{is a float number as the parameter of the underlying exponential distribution}

\item{t}{is a float as the length of time interval.}
}
\description{
If an exponentially distributed variable is modeled as the waiting time before an arrival, the Poisson distributed variable can be modeled as the number of arrivals during a period of time of length t.
}
\seealso{
dugen(), cugen(), brgen(), bigen(), gegen(), expgen(), gagen(), pogen(), nogen() 
related functions: expgen
}
\author{
MohammadAmin Salarkia
}
