% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{point_line_bar}
\alias{point_line_bar}
\title{Allows you to create a point, line or bar graph with your own parameters}
\usage{
point_line_bar(
  data,
  x,
  y,
  geom = "point",
  title = "title",
  x_lab = "x-lab",
  y_lab = "y-lab",
  color = "hotpink",
  alpha = 0.5,
  size = 2,
  theme = theme_minimal()
)
}
\arguments{
\item{data}{the data you want to plot}

\item{x}{the variable you want on the x-axis}

\item{y}{the variable you want on the y-axis}
}
\description{
Allows you to create a point, line or bar graph with your own parameters
}
