% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_process_field_function.R
\name{positive}
\alias{positive}
\title{positive
check data are strictly positive (>0)}
\usage{
positive(dataVector)
}
\arguments{
\item{dataVector}{value vector}
}
\value{
boolean
}
\description{
positive
check data are strictly positive (>0)
}
