% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{power}
\usage{
power(x, n)
}
\arguments{
\item{x}{An integer value( podstawa potegi)}

\item{n}{An integer value (wykladnik potegi)}
}
\description{
Jest to funkcja, która danego x podnosi do potęgi n
}
