% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmmf.R
\name{print.mmf}
\alias{print.mmf}
\title{print}
\usage{
\method{print}{mmf}(x, ...)
}
\arguments{
\item{x}{mmf to handle}

\item{...}{unused}
}
\description{
This function should be called by print(object) and prints out the estimated coefficients.
}
\examples{
 print(testbeta())
}

