% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mychi.R
\name{print.mychi}
\alias{print.mychi}
\title{Print Function for "mychi"}
\usage{
\method{print}{mychi}(x, ...)
}
\arguments{
\item{x}{is a "mychi" class object created by the model helper functions.}

\item{...}{treat like normal print function}
}
\description{
This is a custom print function for the "mychi" class.
}
