% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mylm.R
\name{print.mylm}
\alias{print.mylm}
\title{Print Function for "mylm"}
\usage{
\method{print}{mylm}(x, ...)
}
\arguments{
\item{x}{is a "mylm" class object created by the model helper functions.}

\item{...}{treat like normal print function.}
}
\description{
This is a custom print function for the "mylm" class.
}
