% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.myt.R
\name{print.myt}
\alias{print.myt}
\title{Print Function for "myt"}
\usage{
\method{print}{myt}(x, ...)
}
\arguments{
\item{x}{is a "myt" class object created by the model helper functions.}

\item{...}{treat like normal print function}
}
\description{
This is a custom print function for the "myt" class.
}
