% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project-package.R
\docType{package}
\name{Project-package}
\alias{Project}
\alias{Project-package}
\title{Project: An analytical or development project}
\description{
Represents an analytical or development project, where project is synonymous with a named goal.
}
\author{
\strong{Maintainer}: Stuart Jefferys \email{srjefferys@gmail.com} (\href{https://orcid.org/0000-0001-6259-085X}{ORCID})

}
\keyword{internal}
