% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.add_to_gitignore.R
\name{project.add_to_gitignore}
\alias{project.add_to_gitignore}
\title{project.add_to_gitignore Add a character string or vector entry to the .gitignore file}
\usage{
project.add_to_gitignore(gitignore_additions, .wd = getwd())
}
\arguments{
\item{gitignore_additions}{A character string or character vector of items to add to the .gitignore file}

\item{.wd}{The base working directory of the project. This should contain the .gitignore file being updated (DEFAULT is getwd())}
}
\description{
project.add_to_gitignore Add a character string or vector entry to the .gitignore file
}
\examples{
# For a .gitignore file in the current working directory
wd = getwd()
if(file.exists(glue::glue("{wd}/.gitignore"))){
  project.add_to_gitignore("config.R")
}
}
