% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.copy_config_files.R
\name{project.copy_config_files}
\alias{project.copy_config_files}
\title{Searches through a base folder and looks for config.R files - if found - creates an identical folder name in the output folder and copies the config.R file there}
\usage{
project.copy_config_files(base_folder, output_folder)
}
\arguments{
\item{base_folder}{The folder that contains subfolders to search through}

\item{output_folder}{The folder to store the results in}
}
\description{
Searches through a base folder and looks for config.R files - if found - creates an identical folder name in the output folder and copies the config.R file there
}
