% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.create_gitignore.R
\name{project.create_gitignore}
\alias{project.create_gitignore}
\title{project.create_gitignore Initialize a .gitignore file in the given working directory if it does not already exist}
\usage{
project.create_gitignore(.wd = getwd())
}
\arguments{
\item{.wd}{The base working directory of the project. This should be where the .gitignore file lives (DEFAULT is getwd())}
}
\description{
project.create_gitignore Initialize a .gitignore file in the given working directory if it does not already exist
}
\examples{
# For a .gitignore file in the current working directory
wd = getwd()
project.create_gitignore(wd)
}
