% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.initialize.R
\name{project.initialize}
\alias{project.initialize}
\title{project.initialize Initialize the project}
\usage{
project.initialize(.wd = getwd())
}
\arguments{
\item{.wd}{The base working directory of the project. This should be where the .gitignore file lives (DEFAULT is getwd())}
}
\description{
project.initialize Initialize the project
}
\examples{
project.initialize(getwd())
}
