% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project}
\alias{project}
\title{project}
\format{
\subsection{'project'}{

A data frame with 1000 rows and 5 variables:
\describe{
\item{ID}{ID of the person}
\item{gender}{gender of the person}
\item{height}{height of the person}
\item{weight}{weight of the person}
\item{phys}{amount of physical activity of the person}
...
}
data(project)
}
}
\usage{
project
}
\description{
Dataset for the project
}
\keyword{datasets}
