% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.remove_from_gitignore.R
\name{project.remove_from_gitignore}
\alias{project.remove_from_gitignore}
\title{project.remove_from_gitignore Remove a character string or vector entry to the .gitignore file}
\usage{
project.remove_from_gitignore(gitignore_removals, .wd = getwd())
}
\arguments{
\item{gitignore_removals}{A character string or character vector of items to remove from the .gitignore file}

\item{.wd}{The base working directory of the project. This should contain the .gitignore file being updated (DEFAULT is getwd())}
}
\description{
project.remove_from_gitignore Remove a character string or vector entry to the .gitignore file
}
\examples{
# For a .gitignore file in the current working directory
wd = getwd()
if(file.exists(glue::glue("{wd}/.gitignore"))){
  project.remove_from_gitignore("config.R")
}
}
