% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.set_working_directory.R
\name{project.set_working_directory}
\alias{project.set_working_directory}
\title{An Rstudio interactive UI to set your working directory}
\usage{
project.set_working_directory()
}
\value{
NULL if cancelled or directory not selected; the URL of the directory is selected
}
\description{
An Rstudio interactive UI to set your working directory
}
\examples{
project.set_working_directory()
}
