% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project2022}
\alias{project2022}
\title{Project2022 Dataset}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
\item{\code{ID}}{The ID of each person}
\item{\code{gender}}{The gender of each person}
\item{\code{height}}{The height of each person}
\item{\code{weight}}{The weight of each person}
\item{\code{phys}}{The intensity of physical activity for each person}
}
}
\usage{
project2022
}
\description{
Data sampled from 1000 people.
}
\keyword{datasets}
