% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2.R
\name{q2}
\alias{q2}
\title{Two sample t Test}
\usage{
q2(file_name)
}
\arguments{
\item{file_name}{The name of the excel file that will be read}
}
\value{
A statistic test with the hypothesis, tests assumptions, test statistics, a decision and conclusion.
}
\description{
Runs a two sample t test to see if the average heights of male is the same as females.
}
