% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q3.R
\name{q3}
\alias{q3}
\title{Chi Square Test}
\usage{
q3(file_name)
}
\arguments{
\item{file_name}{The name of the excel file that will be read}
}
\value{
A statistic test with the hypothesis, tests assumptions, test statistics, a decision and conclusion.
}
\description{
Runs a Chi-square statistical test to see whether there is a relationship between the gender and level of physical activity.
}
