% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question1.R
\name{question1}
\alias{question1}
\title{Research question 1}
\usage{
question1(file_name)
}
\arguments{
\item{file_name}{This is the name of the data set}
}
\value{
A statistical test to answer the research question.
}
\description{
This function helps to run the two sample t test in order to answer whether the height and weight are linearly related.
}
