% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question2.R
\name{question2}
\alias{question2}
\title{Research question 2}
\usage{
question2(file_name)
}
\arguments{
\item{file_name}{This is the name of the data set}
}
\value{
A statistical test to answer the research question.
}
\description{
This function helps to run the two sample t test.
}
