% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question3.R
\name{question3}
\alias{question3}
\title{Research question 3}
\usage{
question3(file_name)
}
\arguments{
\item{file_name}{This is the name of the data set}
}
\value{
A statistical test to answer the research question.
}
\description{
Answers whether there any association between gender and the amount of physical activity
}
