% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickview.R
\name{quickview}
\alias{quickview}
\title{Quickly view a variable with \code{id} and \code{visitdate}}
\usage{
quickview(df, key)
}
\arguments{
\item{df}{A data set containing \code{id} and \code{visitdate}. If it doesn't contain those variables, it will only view the \code{key} variable.}

\item{key}{A character string of the name of the variable in \code{df} to view.}
}
\value{
Opens a View() window.
}
\description{
Quickly view a variable with \code{id} and \code{visitdate}
}
