% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rando_color_plot.R
\name{random_color_bar_plot}
\alias{random_color_bar_plot}
\title{Bar plot with random colors}
\usage{
random_color_bar_plot(df, column, title, x_title, y_title)
}
\arguments{
\item{df}{The data frame you want to pull data from.}

\item{column}{The column you want to plot (should be a categorical variable).}

\item{title}{The title of the plot.}

\item{x_title}{The title of the x-axis.}

\item{y_title}{The title of the y-axis.}
}
\value{
A bar graph displaying the frequency of the selected column with random colors.
}
\description{
Bar plot with random colors
}
