% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_airbnb.R
\name{read_airbnb_csv}
\alias{read_airbnb_csv}
\title{Read and Clean Airbnb CSV Data}
\usage{
read_airbnb_csv(
  file_path = system.file("extdata", "airbnb_data.csv", package = "DC_property_data")
)
}
\arguments{
\item{file_path}{The path to the CSV file. Defaults to the package's `inst/extdata/` folder.}
}
\value{
A cleaned dataframe with selected columns.
}
\description{
This function reads a CSV file containing Airbnb data and returns a cleaned dataframe.
}
\examples{
df <- read_airbnb_csv()
}
