% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_real_estate.R
\name{read_real_estate_json}
\alias{read_real_estate_json}
\title{Read and Clean Real Estate JSON Data}
\usage{
read_real_estate_json(
  file_path = system.file("extdata", "washington_dc_real_estate.json", package =
    "DC_property_data")
)
}
\arguments{
\item{file_path}{The path to the JSON file. Defaults to the package's `inst/extdata/` folder.}
}
\value{
A cleaned dataframe with selected columns.
}
\description{
This function reads a JSON file containing real estate data, flattens nested structures, and returns a cleaned dataframe.
}
\examples{
df <- read_real_estate_json()
}
