% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variable.R
\name{rename_variable}
\alias{rename_variable}
\title{Rename variable in column}
\usage{
rename_variable(df, old_name, new_name)
}
\arguments{
\item{df}{the data frame you are processing}

\item{old_name}{the name of the variable you wish to change}

\item{new_name}{the name you want the variable to change to}

\item{column}{the column name where the variable you want to rename is found}
}
\value{
Renamed variable
}
\description{
Rename variable in column
}
