% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{rgenerator}
\alias{rgenerator}
\title{rgenerator}
\usage{
rgenerator()
}
\description{
This function gives you a number from 0 to 2^32-1. It uses xorwow algorithm.
}
\author{
AmirHossein Motameni
}
