% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_anova.R
\name{run_anova}
\alias{run_anova}
\title{Run one-sample ANOVA and t-test}
\usage{
run_anova(df, numeric_column, mu = 0)
}
\arguments{
\item{df}{The data frame containing the numeric column you want to analyze.}

\item{numeric_column}{The name of the numeric column to test (e.g., "Posterior").}

\item{mu}{The hypothesized value to compare the numeric column's mean to. Default is 0.}
}
\value{
An ANOVA summary table
}
\description{
Run one-sample ANOVA and t-test
}
