% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeurQC.R
\name{Seur.QC}
\alias{Seur.QC}
\title{Seur.QC}
\usage{
Seur.QC(obj)
}
\arguments{
\item{obj}{The seurat object created by readSNV}
}
\value{
modified object
}
\description{
This function takes one argument, which is the seurat object representing the SNV data. The function returns a plot visualizing the genes per nucleus of the sample.
}
