% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_buddy.R
\name{study_buddy}
\alias{study_buddy}
\title{Claims study buddy status for two people.}
\usage{
study_buddy(buddy.1, buddy.2, date)
}
\arguments{
\item{buddy.1}{A name for a buddy.}

\item{buddy.2}{A name for a different buddy.}
}
\value{
A statement of buddy status.
}
\description{
Claims study buddy status for two people.
}
\examples{
study_buddy("Jason", "Rick", "2025-01-31")
}
