% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymmf.R
\name{summary.mmf}
\alias{summary.mmf}
\title{summary}
\usage{
\method{summary}{mmf}(object, ...)
}
\arguments{
\item{object}{mmf to handle}

\item{...}{unused}
}
\description{
This function should be called by summary(object) and prints out the estimated coefficients and the standard errors.
}
\examples{
 summary(testbeta())
}

