% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{summary.normal}
\alias{summary.normal}
\title{Summarize content of a normal object; called implicitely.}
\usage{
\method{summary}{normal}(object, ...)
}
\arguments{
\item{object}{normal object}
}
\description{
Summarize content of a normal object; called implicitely.
}
\examples{

# Load iris dataset 
data('iris')

summary(normal(iris[1:25,], significance=5))

}
