% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_function.R
\name{tabulate_function}
\alias{tabulate_function}
\title{Tabulate variables into a single df}
\usage{
tabulate_function(df, variables)
}
\arguments{
\item{df}{Any dataframe that contains variables to tabulate}

\item{variables}{Vector of variables available in the df to tabulate}
}
\value{
Dataframe with tabulations for each variable.
}
\description{
Tabulate variables into a single df
}
\examples{
tabulate_function(df = iris, variables = c("Sepal.Length", "Sepal.Width"))
}
