% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1.R
\name{test1}
\alias{test1}
\title{Test 1-lm test}
\arguments{
\item{file_name}{is the dataset that will be used for this test}
}
\value{
Returns hypothesis, assumptions, statistical analysis results, decision, conclusion
}
\description{
This function is useful to conduct a lm test
}
\examples{
data(project)
file_name <- project
test1(file_name)
}
