% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2.R
\name{test2}
\alias{test2}
\title{Test 2- t test}
\arguments{
\item{file_name}{is the dataset that will be used for this test}
}
\value{
Returns hypothesis, assumptions, statistical analysis results, decision, conclusion
}
\description{
This function is useful to conduct a t test
}
\examples{
data(project)
file_name <- project
test2(file_name)
}
