% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test3.R
\name{test3}
\alias{test3}
\title{Test 3-chi test}
\arguments{
\item{file_name}{is the dataset that will be used for this test}
}
\value{
Returns hypothesis, assumptions, statistical analysis results, decision, conclusion
}
\description{
This function is useful to conduct a chi test
}
\examples{
data(project)
file_name <- project
test3(file_name)
}
