% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitbeta.R
\name{testbeta}
\alias{testbeta}
\title{testbeta}
\usage{
testbeta()
}
\description{
This function generates data from a beta distribution with parameters 5 and 20. Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}
\examples{
mmfit = testbeta()
}

