% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitdiscretepowerlaw.R
\name{testdiscretepowerlaw}
\alias{testdiscretepowerlaw}
\title{testdiscretepowerlaw}
\usage{
testdiscretepowerlaw()
}
\description{
This function generates data from a discrete power law distribution with gamma=2.6.
 Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}

