% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitmixtwoexp.R
\name{testexpmix}
\alias{testexpmix}
\title{testexpmix}
\usage{
testexpmix()
}
\description{
This function generates data from two exponential distribution and then uses mmfit to estimate the parameters and then graphs the data versus the estimated values.
}
\examples{
mmfit = testexpmix()
}

