% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitgamma.R
\name{testgamma}
\alias{testgamma}
\title{testgamma}
\usage{
testgamma()
}
\description{
This function generates data from a gamma distribution with parameters 5 and 1/0.5.
 Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}

