% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitmixtwopoisson.R
\name{testmixtwopoisson}
\alias{testmixtwopoisson}
\title{testmixtwopoisson}
\usage{
testmixtwopoisson()
}
\description{
This function generates data from a mixture of two poisson distributions with mixing ratio=0.3, lambda1=2, lambda2=5.
 Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}

