% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitfunc.R
\name{testmmfitfunc}
\alias{testmmfitfunc}
\title{testmmfitfunc}
\usage{
testmmfitfunc()
}
\description{
This function generates data from a custom implementation of an exponential distribution with lambda=0.5.
 Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}

