% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitpoisson.R
\name{testpoisson}
\alias{testpoisson}
\title{testpoisson}
\usage{
testpoisson()
}
\description{
This function generates data from a poisson distribution with lambda=0.2.
 Then uses mmfit to estimate lambda and graphs the data versus the estimated values.
}
\examples{
mmfit <- testpoisson()
}

