% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitpowerlaw.R
\name{testpowerlaw}
\alias{testpowerlaw}
\title{testpowerlaw}
\usage{
testpowerlaw()
}
\description{
This function generates data from a continuous power law distribution with parameters 2.7.
 Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}

