% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_t.R
\name{ttest_t}
\alias{ttest_t}
\title{Model for T-test}
\usage{
ttest_t(data1)
}
\arguments{
\item{data1}{A dataframe.Use "project2022" dataset for this.}
}
\value{
Returns important values as a list.
}
\description{
This runs the t-test and extracts the important values.
}
