% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_process_field_function.R
\name{uniqueness}
\alias{uniqueness}
\title{uniqueness
check unicity of value in vector}
\usage{
uniqueness(dataVector)
}
\arguments{
\item{dataVector}{value vector}
}
\value{
boolean
}
\description{
uniqueness
check unicity of value in vector
}
