% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes a standard dataframe for league data}
\usage{
uss_make_matches(data_engsoc, .country_name)
}
\arguments{
\item{data_engsoc}{dataframe from engsoccerdata package}

\item{.country_name}{a country: one of spain, italy, germany, holland, england}
}
\value{
dataframe, tibble
}
\description{
Makes a standard dataframe for league data
}
\examples{
uss_make_matches(engsoccerdata::spain,"Spain")
}
