% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvalence.R
\name{valitem}
\alias{ready}
\alias{valitem}
\title{Valence item analysis}
\usage{
valitem(x, posval, negval, valence, na.rm = TRUE)

ready(x, posval, negval, valence, na.rm = TRUE)
}
\arguments{
\item{x}{A valence data frame where the first row contains item IDs and the
first column contains subject IDs.}

\item{posval}{The original value assigned to code output as a positive response.}

\item{negval}{The original value assigned to code output as a negative response.}

\item{valence}{data frame with two columns, the first containing stimuli
names (item), the second containing valence information (val).}
}
\description{
Prepares raw valence rating data for classical item analysis (istudy).
}
\examples{
# Get example data to test function
x <- read.csv("trworddata1p2n.csv", row.names = 1)
valence <- read.csv("valdata.csv")
valitem(x, 1, 2, valence)

}

