% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWrangle.R
\name{wrangle.Files}
\alias{wrangle.Files}
\title{wrangle.Files
Import data table list into R (according source_file description) and wrangle into new data table list following wrangle_parameter_file description.}
\usage{
wrangle.Files(wrangler_parameter, sources_file)
}
\arguments{
\item{sources_file}{name of data frame on which foreign key constraint is checked}

\item{wrangle_parameter}{file of parameter data frame (see format)}
}
\value{
list of data table
}
\description{
wrangle.Files
Import data table list into R (according source_file description) and wrangle into new data table list following wrangle_parameter_file description.
}
