% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWrangle.R
\name{wrangle.TablesList}
\alias{wrangle.TablesList}
\title{wrangle.TablesList
Import data table list into R (according source_file description) and wrangle into new data table list following wrangle_parameter_file description.}
\usage{
wrangle.TablesList(wrangler_parameter, table_list, call_customFunction)
}
\arguments{
\item{wrangle_parameter}{file of parameter data frame (see format)}

\item{sources_file}{name of data frame on which foreign key constraint is checked}
}
\value{
list of data table
}
\description{
wrangle.TablesList
Import data table list into R (according source_file description) and wrangle into new data table list following wrangle_parameter_file description.
}
