% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_chi.R
\name{assumptions_chi}
\alias{assumptions_chi}
\title{Assumptions for Chi-Squared Test}
\usage{
assumptions_chi(data1)
}
\arguments{
\item{data1}{A dataframe.Use "project2022" dataset for this.}
}
\value{
Returns list.
}
\description{
Tabular assumptions for chi-squared test (expected values).
}
