% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_lm.R
\name{assumptions_lm}
\alias{assumptions_lm}
\title{Assumptions for Linear Model}
\usage{
assumptions_lm(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns assumptions as a list.
}
\description{
Provides the assumptions for a linear regression.
}
