% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_t.R
\name{assumptions_t}
\alias{assumptions_t}
\title{Assumptions for T-test}
\usage{
assumptions_t(data1)
}
\arguments{
\item{data1}{A dataframe.Use "project2022" dataset for this.}
}
\value{
Returns assumptions as a list.
}
\description{
Provides the t-test graphical assumptions.
}
