% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_chi.R
\name{chi_chi}
\alias{chi_chi}
\title{Model for Chi-Squared test}
\usage{
chi_chi(data1)
}
\arguments{
\item{data1}{A dataframe.Use "project2022" dataset for this.}
}
\value{
Returns values as a list.
}
\description{
Runs the test and extracts the important values.
}
