% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_function.R
\name{clean_data}
\alias{clean_data}
\title{Replaces blank spaces and omits NAs in data}
\usage{
clean_data(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
clean dataframe
}
\description{
Replaces blank spaces and omits NAs in data
}
