% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crazy_histogram.R
\name{crazy_histogram}
\alias{crazy_histogram}
\title{Histogram plot with random colors}
\usage{
crazy_histogram(df, column, title, x_title, y_title)
}
\arguments{
\item{df}{The data frame you want to pull data from. This should be a data frame containing the column you want to plot.}

\item{column}{The name of the column to plot (should be a numeric variable).}

\item{title}{The title of the plot.}

\item{x_title}{The title of the x-axis.}

\item{y_title}{The title of the y-axis.}
}
\value{
A histogram displaying the frequency of the selected column with random colors.
}
\description{
Histogram plot with random colors
}
