% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{Model for Linear Regression}
\usage{
fit_lm(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns values as a list.
}
\description{
Runs the test and extracts the important values.
}
