% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_chi.R
\name{hypothesis_chi}
\alias{hypothesis_chi}
\title{Hypotheses for Chi-Squared Test}
\usage{
hypothesis_chi(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns hypotheses as a list.
}
\description{
Provides the null and alterantive hypotheses.
}
