% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{logistic_model}
\alias{logistic_model}
\title{logistic model function}
\usage{
logistic_model(data = heart, predictors)
}
\arguments{
\item{predictors}{}
}
\value{
summary of logistic model
}
\description{
logistic model function
}
