% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mywrapper_t.R
\name{mywrapper_t}
\alias{mywrapper_t}
\title{Wrapper Function for T-test}
\usage{
mywrapper_t(data1)
}
\arguments{
\item{data1}{A dataset. Use "project2022" for this project.}
}
\value{
Returns a series of cat/plots/tables.
}
\description{
Calls the helper functions to provide output for the test.
}
