% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmmf.R
\name{plot.mmf}
\alias{plot.mmf}
\title{plot.mmf}
\usage{
\method{plot}{mmf}(obj)
}
\description{
This function should be called by plot(object) and displays two plots side by side.
 The left plot is an empirical cdf with a confidence bound in red and blue around it.
 The right plot a graphing of the parametric and non-parametric data.
}
\examples{
betadist = testbeta()
plot(betadist)
}

