% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generating.R
\name{plot_generating}
\alias{plot_generating}
\title{Plot a boxplot or barplot of variables for inner dataset "heart"}
\usage{
plot_generating(data = heart, variable)
}
\arguments{
\item{data}{using an inner dataset "heart"}

\item{variable}{"}
}
\description{
Plot a boxplot or barplot of variables for inner dataset "heart"
}
