% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{point_line_bar}
\alias{point_line_bar}
\title{Allows you to create a point, line or bar graph with your own parameters}
\usage{
point_line_bar(
  data,
  x,
  y,
  geom = "point",
  title = "title",
  x_lab = "x-lab",
  y_lab = "y-lab",
  color = "hotpink",
  alpha = 0.5,
  size = 2,
  theme = theme_minimal()
)
}
\arguments{
\item{data}{the data you want to plot}

\item{x}{the variable you want on the x-axis}

\item{y}{the variable you want on the y-axis}

\item{geom}{what kind of graph do you want (point, bar or line)}

\item{title}{title of the graph}

\item{x_lab}{x-axis label}

\item{y_lab}{y-axis label}

\item{color}{color you want on your graph}

\item{alpha}{alpha size}

\item{size}{size of graph}

\item{theme}{what theme do you want}
}
\value{
returns graph that user created
}
\description{
Allows you to create a point, line or bar graph with your own parameters
}
