% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_function_prediction.R
\name{predict_regression_model}
\alias{predict_regression_model}
\title{prediction for logistic regression}
\usage{
predict_regression_model(data, predictors, new_data)
}
\arguments{
\item{new_data}{}
}
\value{
probability of new observation's heart attack based on the predictors
}
\description{
prediction for logistic regression
}
