% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmmf.R
\name{print.mmf}
\alias{print.mmf}
\title{print.mmf}
\usage{
\method{print}{mmf}(obj)
}
\arguments{
\item{obj}{mmf to handle}
}
\description{
This function should be called by pint(object) and prints out the estimated coefficients.
}
\examples{
 betadist = testbeta()
 print(betadist)
}

