% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-package.R
\docType{package}
\name{project-package}
\alias{project}
\alias{project-package}
\title{project: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lomidzebeka/project}
  \item \url{https://lomidzebeka.github.io/project/}
  \item Report bugs at \url{https://github.com/lomidzebeka/project/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
