% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.replace_config_files.R
\name{project.replace_config_files}
\alias{project.replace_config_files}
\title{Take in an input_folder with folders that contain config.R files - searches through final_folder and if the subfolder exists then copies config.R to the subfolder in the final_folder}
\usage{
project.replace_config_files(input_folder, final_folder)
}
\arguments{
\item{input_folder}{The folder that contains subfolders with the currently existing config.R files}

\item{final_folder}{The folder that contains the subfolders that you want to copy the config.R files to}
}
\description{
Function partially written with help from Gene.AI
}
