% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project2022}
\alias{project2022}
\title{Data set to run statistical tests}
\format{
A data frame with 1000 rows and 5 variables (columns):
\describe{
\item{ID}{Identification number of an individual person}
\item{gender}{gender of an individual person (i.e., male or female)}
\item{height}{height of an individual person (in cm)}
\item{weight}{weight of an individual person (in kg)}
\item{phys}{physical activity of an individual person}
}
}
\usage{
project2022
}
\description{
A data set which includes ID, gender, height, weight and physical activity
}
\keyword{datasets}
