% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{readSNV}
\alias{readSNV}
\title{readSNV}
\usage{
readSNV(path)
}
\arguments{
\item{path}{The path to the raw SNV data}
}
\value{
The Seurat object representing the data
}
\description{
This function takes one argument, which is the path to the raw SNV data. The function returns a SEURAT object from the data, which is simply the matrix processed into a format friendly to R.
}
