% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_anova.R
\name{run_anova}
\alias{run_anova}
\title{run anova}
\usage{
run_anova(df, numeric_var, group_var)
}
\arguments{
\item{df}{the data frame you want to clean}

\item{num_var}{the name of the numeric variable}

\item{grp_var}{the name of the grouping variable}
}
\value{
An ANOVA summary table
}
\description{
run anova
}
