% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfitmixtwoexp.R
\name{testexpmix}
\alias{testexpmix}
\title{testexpmix}
\usage{
testexpmix()
}
\description{
This function generates data from an exp mixture with mixing ratio=0.5, labda1=1, lambda2=1.
 Then uses mmfit to estimate the parameters and graphs the data versus the estimated values.
}
\examples{
mmfit = testexpmix()
}

