% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_t.R
\name{ttest_t}
\alias{ttest_t}
\title{Model for T-test}
\usage{
ttest_t(project2022)
}
\arguments{
\item{project2022}{dataset}
}
\value{
Returns important values as a list.
}
\description{
This runs the t-test and extracts the important values.
}
